rename marking type single->custom
